Moved TablePaginationActions component to a separate file
